requirements.txt
# api
fastapi>=0.109.1
mangum>=0.17.0
pydantic>=2.6.4
dnspython>=2.7.0
email-validator>=2.2.0

# local running
uvicorn>=0.35.0

# logger
loguru>=0.7.2

# config
pydantic-settings>=2.2.1
dotenv>=0.9.9

# database
supabase>=2.17.0

# authentication
python-multipart>=0.0.20
python-jose>=3.5.0
bcrypt>=4.3.0